Fix DatTime end_of_month calculation for Rails
Using the Holiday gem will override some date caluclation methods in
rails. This if fine for just `Date` objects, however, it will alter the
result of `DateTime.now.end_of_month` greatly as it will not set the
hour, minutes and seconds accordingly and just just the time of the
method call.

This will cause `DateTime.end_of_month` to returd a different end of
month (in terms of hours, minutes and seconds) based on the time of
calling the method and introduces inconsistent data.

This commit introduces an override of the `DateTime` class and uses part
of the calculations for the end_of_month that Rails uses.